backend/requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sounddevice==0.5.2
soundfile==0.13.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
urllib3==2.4.0

# New FastAPI and backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
firebase-admin==6.2.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
